refactor(opentelemetry): inline retry spans into ydb.retries
Move ydb.RunWithRetry / ydb.Try span emission directly into
retry_operation_sync / retry_operation_async in ydb/retries.py, and drop
the short-lived ydb.query._retries shim. Tracing is still no-op by
default, so there is no cost for the table-service callers that share
the same retry loop; we just stop duplicating the retry logic to add
spans.
